Remove Open Sans
It turns out we were loading it but never actually using it because
Bulma's font family line was overriding ours. Helvetica looks better
than the Google Fonts ones that I tried, so let's just stick with that
and save some bandwidth in the process by not loading our own font.